test: add tests for updating token rate cache
